ESLint file renaming
