+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Dependency upgrades
 
+### Fixed
+
+- Rename utility files based on ESLint rule
+
 ## [7.1.0] - 2025-02-16
 
 ### Changed
