+++ src/app/layout.tsx
@@ -7,7 +7,7 @@ import { PropsWithChildren, ReactNode } from 'react';
 import Footer from 'src/components/footer/footer';
 import Header from 'src/components/header/header';
 import { deployURL, protocol } from 'src/helpers/environment';
-import { cn, fullName } from 'src/helpers/utils';
+import { cn, fullName } from 'src/helpers/utilities';
 import './styles/globals.css';
 
 const albert = Albert_Sans({
