+++ src/app/opengraph-image.tsx
@@ -4,7 +4,7 @@ import { ImageResponse } from 'next/og';
 import { readFile } from 'node:fs/promises';
 import path from 'node:path';
 import { getAccentColor, getNeutralColor } from 'src/helpers/colors';
-import { fullName } from 'src/helpers/utils';
+import { fullName } from 'src/helpers/utilities';
 
 const theme = resumeConfig.imageTheme;
