+++ src/components/articles/achievements.tsx
@@ -3,7 +3,7 @@ import { Fragment, ReactNode } from 'react';
 import AchievementItem from 'src/components/articles/achievement-item';
 import Separator from 'src/components/articles/separator';
 import SectionHeading from 'src/components/section-heading/section-heading';
-import { sortedAchievements } from 'src/helpers/utils';
+import { sortedAchievements } from 'src/helpers/utilities';
 
 export default function Achievements(): ReactNode {
   return (
