+++ src/components/section-heading/section-heading.tsx
@@ -1,6 +1,6 @@
 import React, { ReactNode } from 'react';
 import { Heading, HeadingProperties } from 'src/components/heading/heading';
-import { cn } from 'src/helpers/utils';
+import { cn } from 'src/helpers/utilities';
 
 interface SectionHeadingProperties {
   className?: string;
