+++ src/components/star-rating/star-rating.test.tsx
@@ -1,6 +1,6 @@
-import { render } from '@test-utils';
 import { describe, expect, test } from 'bun:test';
 import StarRating from 'src/components/star-rating/star-rating';
+import { render } from 'src/test-utilities';
 
 describe('<StarRating />', () => {
   test('Snapshot', () => {
