+++ src/components/theme-toggle/theme-toggle.test.tsx
@@ -1,6 +1,6 @@
-import { render, screen, userEvent } from '@test-utils';
 import { beforeAll, describe, expect, test } from 'bun:test';
 import { ThemeToggle } from 'src/components/theme-toggle/theme-toggle';
+import { render, screen, userEvent } from 'src/test-utilities';
 
 describe('<Dropdown />', () => {
   beforeAll(() => {
